test it

+++ appveyor.py
@@ -3,14 +3,13 @@ import sys
 import time
 
 import requests
-
-
-def build():
-    pyversion = os.environ.get('TRAVIS_PYTHON_VERSION', "didn't get it")
-    HEADERS = {
+HEADERS = {
         'Content-Type': 'application/json',
         'Authorization': 'Bearer vooe775fjw49jt0iu0d1'
     }
+
+def build():
+    pyversion = os.environ.get('TRAVIS_PYTHON_VERSION', "didn't get it")
     URL = 'https://ci.appveyor.com//api/builds'
     if pyversion == "pypy-5.3.1":
         pr = os.environ.get('TRAVIS_PULL_REQUEST', None)
@@ -39,10 +38,6 @@ def build():
 
 
 def wait():
-    HEADERS = {
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer vooe775fjw49jt0iu0d1'
-    }
     URL = 'https://ci.appveyor.com/api/projects/Eldinnie/python-telegram-bot-6akeh'
     print("checking if Appveyor is still running")
     r = requests.get(url=URL, headers=HEADERS)
@@ -69,29 +64,39 @@ def wait():
                 URL = "https://ci.appveyor.com/api/builds/Eldinnie/python-telegram-bot-6akeh/{}".format(version)
                 r = requests.delete(URL, headers=HEADERS)
                 sys.exit()
+            else:
+                waiting()
         else:
+            try:
+                appveyor_pr = result['build']['pullRequestId']
+            except KeyError:
+                appveyor_pr = ''
             branch = os.environ.get('TRAVIS_BRANCH', None)
             appveyor_branch = result['build']['branch']
-            if branch == appveyor_branch:
+            if branch == appveyor_branch and not appveyor_pr:
                 print("New commit on the same branch, canceling build on Appveyor")
                 version = result['build']['version']
                 URL = "https://ci.appveyor.com/api/builds/Eldinnie/python-telegram-bot-6akeh/{}".format(version)
                 r = requests.delete(URL, headers=HEADERS)
                 sys.exit()
             else:
-                print("Appveyor working, waiting for completion.")
-                while True:
-                    time.sleep(30)
-                    URL = 'https://ci.appveyor.com/api/projects/Eldinnie/python-telegram-bot-6akeh'
-                    r = requests.get(url=URL, headers=HEADERS)
-                    result = r.json()
-                    try:
-                        status = result['build']['status']
-                    except KeyError:
-                        status = ""
-                    if status != 'running':
-                        print('Appveyor is done. Continue')
-                        sys.exit()
+                waiting()
+
+
+def waiting():
+    print("Appveyor working, waiting for completion.")
+    while True:
+        time.sleep(30)
+        URL = 'https://ci.appveyor.com/api/projects/Eldinnie/python-telegram-bot-6akeh'
+        r = requests.get(url=URL, headers=HEADERS)
+        result = r.json()
+        try:
+            status = result['build']['status']
+        except KeyError:
+            status = ""
+        if status != 'running':
+            print('Appveyor is done. Continue')
+            sys.exit()
 
 
 if __name__ == "__main__":
